added code to find breakpoints
Code now plots histograms of the movement data by guild.
Identifies breakpoints by the mean + sd of the data. This may be the
right way to identify the inflection point of continuous data for a
normal distribution. These are not normal distributions, though, and the
data is not technically continuous. Think some more on this.
Then it takes the identified breakpoints and uses them to create MARK
data structures.